auth/requirements.txt
python-dotenv==0.20.0
flask==2.1.1
Flask-SQLAlchemy==2.5.1
isort==5.10.1
pydantic==1.9.0
gunicorn==20.1.0
psycopg2-binary==2.9.1
email-validator==1.1.3
Flask-Security==3.0.0
flask-jwt-extended==4.3.1
flasgger==0.9.5
Werkzeug==2.1.1
marshmallow==3.15.0
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.22.3
Flask-Migrate==3.1.0
Flask-Script==2.0.6